requirements.txt
asgiref==3.2.5
dj-database-url==0.5.0
Django==3.0.3
django-crispy-forms==1.8.1
django-heroku==0.3.1
gunicorn==20.0.4
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==7.0.0
psycopg2==2.7.5
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
